move QueryNode, QueryNodeManager, QueryNodeStartOptions to /node
