+++ substrate-query-framework/index-builder/src/model/QueryEventProcessingPack.ts
@@ -1,10 +1,10 @@
 // @ts-check
 
-import { DatabaseManager, SubstrateEvent } from '.';
+import { DatabaseManager, SubstrateEvent } from '..';
 
 export type QueryEventProcessorResult = void | Promise<void>;
 export type EventHandlerFunc = (db: DatabaseManager, event: SubstrateEvent) => QueryEventProcessorResult
 
-export default interface QueryEventProcessingPack {
+export interface QueryEventProcessingPack {
   [index: string]: EventHandlerFunc;
 }
