+++ substrate-query-framework/index-builder/src/model/index.ts
@@ -1,0 +1,4 @@
+export * from './QueryEvent';
+export * from './QueryEventBlock';
+export * from './QueryEventProcessingPack';
+
